requirements.txt
Django==2.0.1
django-environ==0.4.4
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-invitations==1.9.2
django-allauth==0.34.0
django-favicon==0.1.3
pygsheets==1.1.3
mysqlclient==1.3.12
uwsgi==2.0.15
ijson==2.3
imagehash==4.0
django-extensions==2.0.7
boto3==1.7.48
psycopg2-binary==2.7.5
simplejson
django-cors-headers==2.4.0
